% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TideStations.R
\name{get_reference_station}
\alias{get_reference_station}
\title{Retrieve a subset of information for the selected tide station}
\usage{
get_reference_station(station_code, verbose, harms = MarineTides::harmonics)
}
\arguments{
\item{station_code}{A character field}

\item{verbose}{A boolean requesting additional information be printed to the R console}

\item{harms}{Harmonics data}
}
\value{
A list of text values including \code{station_code, station_type, timezone}. If the tide
station is a subordinate station the \code{station_code} will be that of the harmonic reference
station. The remaining values, \code{timezone} and \code{station_type}, will always belong to
the tide station identified by the \code{station_code} variable, regardless of whether it is
a subordinate or harmonic station.
}
\description{
Retrieves a list of tide station data including \code{timezone} and \code{station_type}
(harmonic or subordinate). If the \code{station_type} is \code{subordinate} it will retrieve
the \code{station_code} for the harmonic reference station, otherwise it will return the same
\code{station_code} that was entered as a variable. The \code{verbose} argument enables
printing additional information to the console window regarding status of tide station
as \code{harmonic} or \code{subordinate}, and how tide predictions are generated for
subordinate stations.
}
