% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TideLevels.R
\name{harmonic_tides}
\alias{harmonic_tides}
\title{Predict tide levels for harmonic stations}
\usage{
harmonic_tides(
  station_code,
  station_info,
  start_date,
  end_date,
  prediction_dts,
  timezone,
  verbose,
  harms
)
}
\arguments{
\item{station_code}{A character field identical to the NOAA Station ID}

\item{station_info}{A list including codes for the reference or subordinate stations}

\item{start_date}{A character field for the start date of the tide prediction time-span}

\item{end_date}{A character field for the end date of the tide prediction time-span}

\item{prediction_dts}{A vector of datetimes for the tide prediction.}

\item{timezone}{Typically the timezone of the \code{tide_station}, but can also be set manually.}

\item{verbose}{A boolean requesting additional information be printed to the R console.}

\item{harms}{Harmonics data}
}
\value{
A dataframe of predicted values. Includes information on the tide station
}
\description{
Given a \code{station_code}, a vector of dates (\code{prediction_dts}), and some additional
information in the \code{station_info} list output by \code{MarineTides::get_reference_station},
the \code{harmonic_tides()} function generates tide predictions for one harmonic station. It is
the core function called by the \code{MarineTides::tide_level()} function.
}
