% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{harmonics}
\alias{harmonics}
\title{Harmonics}
\format{
An object of class \code{list} of length 5.
}
\usage{
harmonics
}
\description{
A set of harmonic data for NOAA CO-OPS tide stations. Also includes year constituents.
}
\details{
All tide station data, including constituent data and subordinate station offsets, were
downloaded from the [NOAA CO-OPS Data API](https://api.tidesandcurrents.noaa.gov/api/prod/)
Year constituent data captures the effects of the 18.6 year orbital cycle on amplitudes
of tidal constituents. Values for years 1700 to 2100 were extracted from the [rtide::harmonics]
data originally processed by David Flater for [XTide](https://flaterco.com/xtide/) and repackaged
for use in the`rtide` package by Joe Thorley of Poisson Consulting Ltd.
}
\keyword{datasets}
