% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarineTides-package.R
\name{MarineTides.package}
\alias{MarineTides.package}
\title{MarineTides}
\description{
Generates tide predictions for harmonic reference stations, and subordinate stations. Uses
harmonic data obtained from NOAA CO-OPS API
}
\author{
Are Strom

Maintainer: Are Strom <are.strom@gmail.com>
}
\keyword{package}
