% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{harmonics}
\alias{harmonics}
\title{Harmonics}
\format{
An object of class \code{list} of length 5.
}
\usage{
harmonics
}
\description{
A set of harmonic data for NOAA CO-OPS tide stations. Also includes node year constituents.
}
\details{
All tide station data, including harmonic constituents and subordinate station offsets, were
downloaded from the \href{https://api.tidesandcurrents.noaa.gov/api/prod/}{NOAA CO-OPS Data API}

Lunar node year data captures the effects of the 18.6 year lunar orbital cycle on amplitudes
and phase of tidal constituents. Values for years 1700 to 2100 were extracted from the
\href{https://github.com/poissonconsulting/rtide/tree/main/data}{rtide::harmonics}
dataset originally processed by David Flater for \href{https://flaterco.com/xtide/}{XTide} and repackaged
for use in the \code{rtide} package by Joe Thorley of Poisson Consulting Ltd.

The names of some constituents needed to be updated to match naming conventions used by NOAA CO-OPS. The
\href{https://cran.r-project.org/web/packages/TideHarmonics/TideHarmonics.pdf}{`TideHarmonics::harmonics`}
dataset was used to aid in making translations between naming conventions. Documentation associated
with the \code{congen} program at \href{https://flaterco.com/files/xtide/congen_input.txt}{`XTide`}
served as an additional resource to update constituent names.
}
\keyword{datasets}
