% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MarineTides-package.R
\docType{package}
\name{MarineTides.package}
\alias{MarineTides}
\alias{MarineTides-package}
\alias{MarineTides.package}
\title{MarineTides}
\description{
Generates tide predictions for harmonic reference stations, and subordinate stations. Uses
harmonic data obtained from NOAA CO-OPS API
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/arestrom/MarineTides}
  \item Report bugs at \url{https://github.com/arestrom/MarineTides/issues}
}

}
\author{
Are Strom

Maintainer: Are Strom <are.strom@gmail.com>
}
\keyword{internal}
